# Labels: simpler/faster stringlabels encoding