@bboreham 2025-03-01
Every branch slows it down. This is why I went for a very simple scheme. 
However it isn’t that great on AMD64. Maybe some analysis can improve it.