@machine424 2025-03-03
I missed that, I even re-ran the benchmarks, but I have an Arm x)
It's true that we end up with either 1 or 4 memory accesses, with no in-between, but with less logic/branching and that actually suits Arm better. So yes, maybe there are some changes we could make to optimize for Amd as well.
I'll take a look as well when I have some time.