@machine424 2025-03-10
Found myself looking at the assembly for regressions, apparently hoisting (I thought `gc` was able to do that) `index` in `decodeSize` 

```
 func decodeSize(data string, index int) (int, int) {
        b := data[index]
        index++
        if b < 255 {
                return int(b), index
        }
        ...
```

helps speed up some of them, the change also makes that 1-byte case "identical" with the code in main.

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/model/labels
cpu: AMD EPYC 7R32
                                                │ prometheus-main/main.txt │        prometheus-pr/pr.txt         │ prometheus-pr-hoisting/pr-hoisting.txt │
                                                │          sec/op          │    sec/op     vs base               │     sec/op       vs base               │
Labels_Get/with_5_labels/first_label/get-8                     7.801n ± 1%    7.427n ± 1%   -4.79% (p=0.002 n=6)      7.744n ±  1%        ~ (p=0.065 n=6)
Labels_Get/with_5_labels/first_label/has-8                     5.896n ± 1%    6.806n ± 1%  +15.45% (p=0.002 n=6)      5.923n ±  1%        ~ (p=0.394 n=6)
Labels_Get/with_5_labels/middle_label/get-8                    11.03n ± 2%    11.17n ± 0%        ~ (p=0.071 n=6)      11.01n ±  2%        ~ (p=0.903 n=6)
Labels_Get/with_5_labels/middle_label/has-8                    8.940n ± 2%   10.530n ± 3%  +17.79% (p=0.002 n=6)      8.908n ±  1%        ~ (p=0.589 n=6)
Labels_Get/with_5_labels/last_label/get-8                      16.43n ± 1%    16.64n ± 0%   +1.25% (p=0.002 n=6)      16.16n ±  0%   -1.64% (p=0.002 n=6)
Labels_Get/with_5_labels/last_label/has-8                      13.40n ± 1%    14.62n ± 0%   +9.06% (p=0.002 n=6)      13.35n ±  1%   -0.45% (p=0.032 n=6)
Labels_Get/with_5_labels/not-found_label/get-8                 6.706n ± 1%    7.380n ± 0%  +10.04% (p=0.002 n=6)      7.423n ±  0%  +10.68% (p=0.002 n=6)
Labels_Get/with_5_labels/not-found_label/has-8                 6.831n ± 2%    7.785n ± 8%  +13.97% (p=0.002 n=6)      6.685n ±  0%   -2.14% (p=0.002 n=6)
Labels_Get/with_10_labels/first_label/get-8                    7.731n ± 1%    7.447n ± 0%   -3.67% (p=0.002 n=6)      7.769n ±  0%        ~ (p=0.558 n=6)
Labels_Get/with_10_labels/first_label/has-8                    5.917n ± 2%    6.832n ± 4%  +15.47% (p=0.002 n=6)      6.098n ±  6%        ~ (p=0.310 n=6)
Labels_Get/with_10_labels/middle_label/get-8                   19.33n ± 0%    19.38n ± 0%   +0.23% (p=0.011 n=6)      18.45n ±  1%   -4.55% (p=0.002 n=6)
Labels_Get/with_10_labels/middle_label/has-8                   15.92n ± 1%    17.25n ± 0%   +8.29% (p=0.002 n=6)      15.88n ±  0%        ~ (p=0.134 n=6)
Labels_Get/with_10_labels/last_label/get-8                     31.89n ± 0%    31.74n ± 0%   -0.45% (p=0.002 n=6)      31.67n ±  0%   -0.69% (p=0.002 n=6)
Labels_Get/with_10_labels/last_label/has-8                     28.34n ± 0%    28.83n ± 0%   +1.71% (p=0.002 n=6)      28.28n ±  0%        ~ (p=0.054 n=6)
Labels_Get/with_10_labels/not-found_label/get-8                6.796n ± 2%    7.413n ± 1%   +9.09% (p=0.002 n=6)      7.417n ±  0%   +9.15% (p=0.002 n=6)
Labels_Get/with_10_labels/not-found_label/has-8                6.809n ± 1%    7.743n ± 1%  +13.72% (p=0.002 n=6)      6.668n ±  0%   -2.06% (p=0.002 n=6)
Labels_Get/with_30_labels/first_label/get-8                    7.763n ± 7%    7.414n ± 0%   -4.48% (p=0.002 n=6)      7.731n ±  0%        ~ (p=0.180 n=6)
Labels_Get/with_30_labels/first_label/has-8                    5.893n ± 0%    6.795n ± 2%  +15.32% (p=0.002 n=6)      6.115n ± 19%   +3.78% (p=0.041 n=6)
Labels_Get/with_30_labels/middle_label/get-8                   50.14n ± 0%    49.94n ± 0%   -0.40% (p=0.002 n=6)      49.88n ±  0%   -0.51% (p=0.002 n=6)
Labels_Get/with_30_labels/middle_label/has-8                   46.62n ± 0%    47.44n ± 0%   +1.77% (p=0.002 n=6)      46.57n ±  0%   -0.11% (p=0.032 n=6)
Labels_Get/with_30_labels/last_label/get-8                     93.72n ± 0%    93.41n ± 0%   -0.33% (p=0.002 n=6)      93.48n ±  0%   -0.26% (p=0.002 n=6)
Labels_Get/with_30_labels/last_label/has-8                     90.26n ± 0%    90.89n ± 0%   +0.70% (p=0.002 n=6)      90.17n ±  0%   -0.09% (p=0.024 n=6)
Labels_Get/with_30_labels/not-found_label/get-8                6.716n ± 1%    7.377n ± 4%   +9.83% (p=0.002 n=6)      7.411n ±  1%  +10.34% (p=0.002 n=6)
Labels_Get/with_30_labels/not-found_label/has-8                6.800n ± 1%    7.721n ± 0%  +13.54% (p=0.002 n=6)      6.699n ±  1%   -1.49% (p=0.002 n=6)
Labels_Equals/equal-8                                          5.267n ± 0%    5.247n ± 0%        ~ (p=0.065 n=6)      5.255n ±  0%        ~ (p=0.167 n=6)
Labels_Equals/not_equal-8                                     0.6199n ± 0%   0.6215n ± 1%   +0.26% (p=0.039 n=6)     0.6200n ±  1%        ~ (p=1.000 n=6)
Labels_Equals/different_sizes-8                               0.6205n ± 1%   0.6193n ± 1%        ~ (p=0.589 n=6)     0.6179n ±  1%   -0.42% (p=0.035 n=6)
Labels_Equals/lots-8                                           5.244n ± 0%    5.255n ± 0%   +0.22% (p=0.037 n=6)      5.240n ±  0%        ~ (p=0.329 n=6)
Labels_Equals/real_long_equal-8                                12.49n ± 1%    12.67n ± 2%   +1.44% (p=0.026 n=6)      12.55n ±  1%        ~ (p=0.368 n=6)
Labels_Equals/real_long_different_end-8                        9.410n ± 2%    9.460n ± 3%        ~ (p=0.240 n=6)      9.413n ±  0%        ~ (p=0.416 n=6)
Labels_Compare/equal-8                                         9.245n ± 4%    9.312n ± 2%        ~ (p=0.485 n=6)      9.008n ±  3%        ~ (p=0.065 n=6)
Labels_Compare/not_equal-8                                     20.88n ± 9%    18.47n ± 0%  -11.56% (p=0.002 n=6)      19.35n ±  5%   -7.35% (p=0.002 n=6)
Labels_Compare/different_sizes-8                               6.993n ± 3%    7.044n ± 0%        ~ (p=0.372 n=6)      7.019n ±  1%        ~ (p=0.589 n=6)
Labels_Compare/lots-8                                          40.14n ± 1%    50.34n ± 0%  +25.38% (p=0.002 n=6)      39.43n ±  0%   -1.79% (p=0.002 n=6)
Labels_Compare/real_long_equal-8                               22.41n ± 3%    22.37n ± 2%        ~ (p=0.699 n=6)      22.75n ±  2%   +1.56% (p=0.039 n=6)
Labels_Compare/real_long_different_end-8                       48.12n ± 0%    46.35n ± 0%   -3.69% (p=0.002 n=6)      46.29n ±  0%   -3.81% (p=0.002 n=6)
Labels_Hash/typical_labels_under_1KB-8                         72.49n ± 0%    72.33n ± 0%        ~ (p=0.056 n=6)      72.59n ±  0%   +0.14% (p=0.009 n=6)
Labels_Hash/bigger_labels_over_1KB-8                           88.73n ± 0%    88.44n ± 0%   -0.34% (p=0.041 n=6)      88.81n ±  0%        ~ (p=0.076 n=6)
Labels_Hash/extremely_large_label_value_10MB-8                 837.7µ ± 0%    827.8µ ± 0%   -1.18% (p=0.002 n=6)      831.3µ ±  0%   -0.76% (p=0.002 n=6)
Builder-8                                                      565.2n ± 0%    500.3n ± 0%  -11.47% (p=0.002 n=6)      527.4n ±  0%   -6.70% (p=0.002 n=6)
Labels_Copy-8                                                  69.42n ± 1%    67.78n ± 2%   -2.36% (p=0.015 n=6)      68.11n ±  1%   -1.89% (p=0.009 n=6)
geomean
```

Still `not-found_label/get` is intriguing, I'll look at that.

Note that only the 1-byte case in `decodeSize` is exercised in the benchmarks above, we should cover the rest.

I also need to run this on ARM.