@bboreham 2025-04-24
Added @machine424's change which makes a decent improvement on amd64.
I found that reversing the conditions in `decodeSize` stops arm64 benchmarking slower.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Apple M2
                                                │  before3.txt  │              after3.txt              │
                                                │    sec/op     │    sec/op      vs base               │
Labels_Get/with_5_labels/first_label/get-8         5.317n ±  1%    4.445n ±  2%  -16.41% (p=0.002 n=6)
Labels_Get/with_5_labels/first_label/has-8         4.423n ±  0%    4.214n ±  2%   -4.73% (p=0.002 n=6)
Labels_Get/with_5_labels/middle_label/get-8        8.264n ±  1%    6.269n ±  1%  -24.14% (p=0.002 n=6)
Labels_Get/with_5_labels/middle_label/has-8        7.381n ±  0%    5.971n ±  1%  -19.10% (p=0.002 n=6)
Labels_Get/with_5_labels/last_label/get-8         10.925n ±  4%    8.883n ±  5%  -18.70% (p=0.002 n=6)
Labels_Get/with_5_labels/last_label/has-8         10.050n ±  1%    8.050n ±  4%  -19.90% (p=0.002 n=6)
Labels_Get/with_5_labels/not-found_label/get-8     6.275n ±  1%    5.181n ±  1%  -17.43% (p=0.002 n=6)
Labels_Get/with_5_labels/not-found_label/has-8     5.364n ± 12%    5.429n ±  0%        ~ (p=0.394 n=6)
Labels_Get/with_10_labels/first_label/get-8        5.314n ±  2%    4.429n ±  4%  -16.64% (p=0.002 n=6)
Labels_Get/with_10_labels/first_label/has-8        4.447n ±  1%    4.141n ±  1%   -6.89% (p=0.002 n=6)
Labels_Get/with_10_labels/middle_label/get-8       12.72n ±  2%    10.82n ±  4%  -15.01% (p=0.002 n=6)
Labels_Get/with_10_labels/middle_label/has-8      11.885n ±  7%    9.804n ±  1%  -17.51% (p=0.002 n=6)
Labels_Get/with_10_labels/last_label/get-8         20.00n ±  7%    18.41n ±  3%   -8.00% (p=0.002 n=6)
Labels_Get/with_10_labels/last_label/has-8         18.62n ±  4%    17.16n ±  1%   -7.82% (p=0.002 n=6)
Labels_Get/with_10_labels/not-found_label/get-8    6.306n ±  7%    5.129n ± 13%  -18.66% (p=0.002 n=6)
Labels_Get/with_10_labels/not-found_label/has-8    5.341n ±  4%    5.468n ±  4%   +2.39% (p=0.041 n=6)
Labels_Get/with_30_labels/first_label/get-8        5.500n ±  9%    4.431n ±  0%  -19.44% (p=0.002 n=6)
Labels_Get/with_30_labels/first_label/has-8        4.491n ± 10%    4.136n ±  0%   -7.89% (p=0.002 n=6)
Labels_Get/with_30_labels/middle_label/get-8       37.44n ±  3%    36.41n ±  0%   -2.72% (p=0.002 n=6)
Labels_Get/with_30_labels/middle_label/has-8       35.48n ±  4%    34.65n ±  4%   -2.33% (p=0.035 n=6)
Labels_Get/with_30_labels/last_label/get-8         77.73n ±  1%    76.51n ±  1%   -1.58% (p=0.002 n=6)
Labels_Get/with_30_labels/last_label/has-8         75.49n ±  3%    74.13n ± 14%        ~ (p=0.061 n=6)
Labels_Get/with_30_labels/not-found_label/get-8    6.272n ±  3%    5.148n ±  2%  -17.91% (p=0.002 n=6)
Labels_Get/with_30_labels/not-found_label/has-8    5.323n ±  0%    5.399n ±  2%   +1.41% (p=0.002 n=6)
Labels_Equals/equal-8                              2.953n ±  2%    2.956n ±  1%        ~ (p=0.446 n=6)
Labels_Equals/not_equal-8                         0.3141n ±  1%   0.3136n ±  0%        ~ (p=0.154 n=6)
Labels_Equals/different_sizes-8                   0.3136n ±  0%   0.3136n ±  0%        ~ (p=0.652 n=6)
Labels_Equals/lots-8                               2.661n ±  1%    2.659n ±  1%        ~ (p=0.675 n=6)
Labels_Equals/real_long_equal-8                    7.973n ±  0%    8.015n ±  7%        ~ (p=0.180 n=6)
Labels_Equals/real_long_different_end-8            6.800n ±  3%    6.794n ±  4%        ~ (p=0.974 n=6)
Labels_Compare/equal-8                             6.498n ±  1%    6.506n ±  1%        ~ (p=0.589 n=6)
Labels_Compare/not_equal-8                         13.59n ±  4%    11.76n ±  4%  -13.46% (p=0.002 n=6)
Labels_Compare/different_sizes-8                   5.484n ±  1%    5.676n ±  1%   +3.52% (p=0.002 n=6)
Labels_Compare/lots-8                              26.52n ±  1%    23.21n ±  1%  -12.52% (p=0.002 n=6)
Labels_Compare/real_long_equal-8                   16.78n ±  1%    16.79n ± 10%        ~ (p=0.193 n=6)
Labels_Compare/real_long_different_end-8           28.31n ±  1%    25.40n ±  0%  -10.29% (p=0.002 n=6)
Labels_Hash/typical_labels_under_1KB-8             54.87n ±  2%    54.55n ±  4%        ~ (p=0.132 n=6)
Labels_Hash/bigger_labels_over_1KB-8               66.97n ±  1%    66.75n ±  0%        ~ (p=0.102 n=6)
Labels_Hash/extremely_large_label_value_10MB-8     676.3µ ±  0%    675.6µ ±  1%        ~ (p=0.818 n=6)
Builder-8                                          251.8n ±  4%    235.9n ±  0%   -6.33% (p=0.002 n=6)
Labels_Copy-8                                      27.91n ±  3%    27.76n ±  2%        ~ (p=0.104 n=6)
geomean                                            13.12n          12.11n         -7.68%
```